+++ tests/job_test.py
@@ -1,4 +1,6 @@
-from beaker import Beaker, CurrentJobStatus, JobKind
+import time
+
+from beaker import Beaker, CurrentJobStatus, ExperimentSpec, JobKind, TaskSpec
 
 
 def test_job_get(client: Beaker, hello_world_job_id):
@@ -7,3 +9,22 @@ def test_job_get(client: Beaker, hello_world_job_id):
     assert job.status.current == CurrentJobStatus.finalized
     assert job.kind == JobKind.execution
     assert job.to_json()["kind"] == "execution"
+
+
+def test_job_stop_and_finalize(client: Beaker, experiment_name: str, beaker_cluster_name: str):
+    start = time.time()
+    spec = ExperimentSpec().with_task(
+        TaskSpec.new(
+            "main",
+            beaker_cluster_name,
+            docker_image="hello-world",
+        ),
+    )
+    experiment = client.experiment.create(experiment_name, spec)
+    while not experiment.jobs:
+        experiment = client.experiment.get(experiment.id)
+        if time.time() - start > 360:
+            raise TimeoutError
+    client.job.stop(experiment.jobs[0])
+    job = client.job.finalize(experiment.jobs[0])
+    assert job.status.current == CurrentJobStatus.finalized